The abstract class Shape and the methods getCircumference and getArea in the class have been modified to throw a ShapeException.
The methods getCircumference and getArea in the classes circle and
rectangle catch the thrown Custom Exception.
In addition the class rectangle and its getHeight and getWidth methods
catch the thrown Custom Exception. Also the class circle and its method
getRadius catch the thrown Custom Exception.
Finally, the main class the tests the application has been changed to
test the new custom exception class.