Note: a lot has happened in the project. This is now the first Alpha version 1.0.
DrawingPanel:
This class, was an inner class, has been move from the class
MyDrawingArea. The move was motivated by the need to add additional
logic. New additional mouse motion events have been added. Logic has
been added to a mouse motion when either a button or not is pressed.
Additional logic has also been added when a mouse button is either
pushed or released.
MyToolRow:
This class has been heavily re-factored. This class, formally an inner
class in MyDrawingArea, has been move to separate class called
MyToolRow. The move was motivated by the need to add additional logic.
Replaced both coloursEnum enum and colours HashMap with the enum class
ToolRowEnum. ToolRowEnum has additional logic and is now more compact
and easier to maintain. Replaced the drawingTools array with the
DrawingToolsEnum Enum class. This will make the class more usable.
Replaced the array drawingTools with an enum class DrawingToolsEnum.
Added additional logic to the JComboBox drawingToolComboBox. When a
selection is made all the listeners to this source will be informed that
a selection has been made. A new empty shape is created and sent to the
listeners.
MyMenuBar:
Added additional logic to the menu items. In addition, both the
MBTextEnum and the Dictionary Text have been replaced with an Enum class
called MenuBarItemEnum. This makes for a cleaner code style. Also the
Enum class allows for additional logic at a later date. In the
actionPerfromed method, all the “if then logic” has been replaced with a
Command Design Pattern. This pattern allows for additional logic to be
added to classes that implement the different commands from the menu.
JPaintFrame:
The class MyDrawingArea has been removed from the application. It was
broken down into 2 new classes, DrawingPanel and MyToolRow. The
motivation for this change is that the menu tool bars logic will be
developed.
The wiring for event handling has been modified.
Shape:
Added a copy and clone constructors. Since this is an abstract class all
derived sub classes must also add copy and clone constructors. Also
included is an isValid method that checks if the shape has a colour and
that all points are not null. Two new overloaded methods have been added
to set the starting point of the shape.

Phew!