Class Step8, add additional documentation on what has been added to this step.
Class FileHandler. This class has been further developed to contain the
logic for loading and saving drawings object. The files that are created
and used for storing drawing objects are formatted text files. The
structure of the save data is as follows.
Name of the art piece, Title.
Name of the artist, Author.
Type of Shape, start x point, start y point, second x point, second y
point, colour.
There maybe 0 or many shapes in the drawing object.